+++ scripts/verify-snaps.ts
@@ -1,4 +1,6 @@
-import { detectSnapLocation } from '@metamask/snaps-controllers/dist/snaps/location';
+import { detectSnapLocation, fetchSnap } from '@metamask/snaps-controllers';
+import type { SnapId } from '@metamask/snaps-sdk';
+import { getLocalizedSnapManifest } from '@metamask/snaps-utils';
 import { assertIsSemVerVersion } from '@metamask/utils';
 import deepEqual from 'fast-deep-equal';
 import semver from 'semver/preload';
@@ -32,9 +34,23 @@ async function verifySnapVersion(
   const location = detectSnapLocation(snap.id, {
     versionRange: version as any,
   });
-  const { result: manifest } = await location.manifest();
 
-  if (latest && snap.metadata.name !== manifest.proposedName) {
+  // This will throw if the snap checksum is invalid etc
+  const fetchedSnap = await fetchSnap(snap.id as SnapId, location);
+
+  const manifest = fetchedSnap.manifest.result;
+  const validatedLocalizationFiles = fetchedSnap.localizationFiles.map(
+    (file) => file.result,
+  );
+
+  // For now, just validate the English translation
+  const localizedManifest = getLocalizedSnapManifest(
+    manifest,
+    'en',
+    validatedLocalizationFiles,
+  );
+
+  if (latest && snap.metadata.name !== localizedManifest.proposedName) {
     throw new Error(
       // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
       `Proposed name for "${snap.id}@${version}" does not match the metadata name in the registry. Expected "${manifest.proposedName}" (proposed name), got "${snap.metadata.name}" (registry metadata name).`,