+++ CHANGELOG.md
@@ -6,6 +6,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.0.1]
+### Changed
+- Bump `@metamask/utils` from `8.2.1` to `8.3.0` ([#374](https://github.com/MetaMask/snaps-registry/pull/374))
+
+### Fixed
+- Fix registry signature validation ([#471](https://github.com/MetaMask/snaps-registry/pull/471))
+
+## [3.0.0]
+### Changed
+- **BREAKING:** Bump minimum Node.js version to `^18.16.0` ([#302](https://github.com/MetaMask/snaps-registry/pull/302))
+- **BREAKING:** Use `noble-hashes` and `noble-curves` ([#315](https://github.com/MetaMask/snaps-registry/pull/315))
+  - This is breaking as the exported `verify` function is no longer asynchronous.
+  - This also improves compatibility with React Native.
+- Add more metadata fields ([#316](https://github.com/MetaMask/snaps-registry/pull/316))
+
+## [2.1.1]
+### Fixed
+- License package under MIT / Apache 2.0 dual license ([#272](https://github.com/MetaMask/snaps-registry/pull/272))
+
+## [2.1.0]
+### Added
+- Add hidden field to metadata ([#226](https://github.com/MetaMask/snaps-registry/pull/226))
+
 ## [2.0.0]
 ### Changed
 - **BREAKING**: Snap IDs are now required to start with the prefix `npm:` ([#75](https://github.com/MetaMask/snaps-registry/pull/75))
@@ -39,7 +62,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release
 
-[Unreleased]: https://github.com/MetaMask/snaps-registry/compare/v2.0.0...HEAD
+[Unreleased]: https://github.com/MetaMask/snaps-registry/compare/v3.0.1...HEAD
+[3.0.1]: https://github.com/MetaMask/snaps-registry/compare/v3.0.0...v3.0.1
+[3.0.0]: https://github.com/MetaMask/snaps-registry/compare/v2.1.1...v3.0.0
+[2.1.1]: https://github.com/MetaMask/snaps-registry/compare/v2.1.0...v2.1.1
+[2.1.0]: https://github.com/MetaMask/snaps-registry/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/MetaMask/snaps-registry/compare/v1.2.2...v2.0.0
 [1.2.2]: https://github.com/MetaMask/snaps-registry/compare/v1.2.1...v1.2.2
 [1.2.1]: https://github.com/MetaMask/snaps-registry/compare/v1.2.0...v1.2.1